Modified crontab installation to be more efficient
Previously, installing a crontab used meant writing the old config to a file on disk, writing to the end of that file, telling cron to use that file as the new crontab, then deleting all the temporarily created files/folder.

Now it reads the old crontab into a variable, modifies that variable, then installs the new crontab by feeding into the stdin of `crontab`. It's much shorter and cleaner, and results in less disk IO.